Refactor cmd command
